pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
pydantic-settings = "^2.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.15.0"
types-python-jose = "^3.4.0.20250224"
types-passlib = "^1.7.7.20241221"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
